% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{az_fmt_batch}
\alias{az_fmt_batch}
\title{Format a Batch of Addresses for Geocoding with Azure Maps}
\usage{
az_fmt_batch(df, address, limit = 1)
}
\arguments{
\item{df}{A data frame containing addresses.}

\item{address}{A column in \code{df} containing the addresses to geocode.}

\item{limit}{An integer. The maximum number of results to return.}
}
\value{
JSON. A formatted JSON string ready for use with the Azure Maps API.
}
\description{
A function to format a batch of addresses for geocoding using Azure Maps API.
}
