% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{az_geocode_batch}
\alias{az_geocode_batch}
\title{Geocode a Batch of Addresses with Azure Maps}
\usage{
az_geocode_batch(batch_json)
}
\arguments{
\item{batch_json}{A JSON. A JSON string formatted for use with the Azure Maps API.}
}
\value{
JSON. The response from the Azure Maps API.
}
\description{
A function to geocode a batch of addresses using Azure Maps API.
}
