% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{az_geocode_file}
\alias{az_geocode_file}
\title{Geocode Addresses from a JSON File with Azure Maps}
\usage{
az_geocode_file(batch_json)
}
\arguments{
\item{batch_json}{File path. Path to the JSON file containing addresses to geocode.}
}
\value{
JSON. The response from the Azure Maps API.
}
\description{
A function to geocode addresses from a JSON file using Azure Maps API.
}
