% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{azm_fmt_rev_batch}
\alias{azm_fmt_rev_batch}
\title{Format a Batch of Coordinates for Reverse Geocoding with Azure Maps}
\usage{
azm_fmt_rev_batch(df, lat, lon, limit = 1)
}
\arguments{
\item{df}{A data frame containing coordinates}

\item{lat}{A column in \code{df} containing the latitude to geocode.}

\item{lon}{A column in \code{df} containing the longitude to geocode.}

\item{limit}{An integer. The maximum number of results to return.}
}
\value{
JSON. A formatted JSON string ready for use with the Azure Maps API.
}
\description{
A function to format a batch of coordinates for geocoding using Azure Maps API.
}
