% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{azm_geocode}
\alias{azm_geocode}
\title{Geocode Addresses with the AZM Geocoding Service}
\usage{
azm_geocode(df, address, type = "position", .test_batch = FALSE)
}
\arguments{
\item{df}{A data frame containing the addresses to be geocoded.}

\item{address}{A string indicating the name of the column in \code{df}
containing the addresses to be geocoded.}

\item{type}{A string indicating the type of geocoding to be performed.
Must be one of "position", "address", or "matchConfidence".}

\item{.test_batch}{A logical. If \code{TRUE}, the function will run in test mode,
processing only a small subset of the data.}
}
\value{
A data frame containing the original data plus geocoding results.
}
\description{
This function interacts with the AZM geocoding service to
geocode addresses in a data frame, handling batching and type selection.
}
