% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{azm_geocode_batch}
\alias{azm_geocode_batch}
\title{Geocode Addresses with Azure Maps}
\usage{
azm_geocode_batch(df, address, type = "position")
}
\arguments{
\item{df}{A data frame containing the addresses to be geocoded.}

\item{address}{A character vector specifying the address column(s) in \code{df}.}

\item{type}{A character string indicating the type of data to return.
Must be one of "position", "address", "matchConfidence", or "all".
Default is "position".}
}
\value{
A data frame containing the original data along with the requested
geocoding information.
}
\description{
Utilize Azure Maps services to geocode addresses, returning either positional
information, address details, confidence scores, or all available details.
}
