% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azm_url.R
\docType{data}
\name{azm_maps_base_url}
\alias{azm_maps_base_url}
\title{Azure Maps API Base URL Component}
\format{
An object of class \code{character} of length 1.
}
\usage{
azm_maps_base_url
}
\description{
A character string storing the base URL component for the Azure Maps API.
}
\keyword{datasets}
