% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{azm_request_batch}
\alias{azm_request_batch}
\title{Request a Batch of Addresses from Azure Maps}
\usage{
azm_request_batch(batch_json)
}
\arguments{
\item{batch_json}{A JSON. A JSON string formatted for use with the Azure Maps API.}
}
\value{
JSON. The response from the Azure Maps API.
}
\description{
A function to request a batch of addresses using Azure Maps API.
}
