% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azm_url.R
\docType{data}
\name{azm_rev_batch_url}
\alias{azm_rev_batch_url}
\title{Azure Maps API Reverse Batch URL Component}
\format{
An object of class \code{character} of length 1.
}
\usage{
azm_rev_batch_url
}
\description{
A character string storing the URL component for the Azure Maps API.
}
\keyword{datasets}
