% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{azm_rev_geocode}
\alias{azm_rev_geocode}
\title{Reverse Geocode Addresses with the AZM Geocoding Service}
\usage{
azm_rev_geocode(df, lat, lon, type = "address", .test_batch = FALSE)
}
\arguments{
\item{df}{A data frame containing the addresses to be geocoded.}

\item{lat}{A column in \code{df} containing the latitude to geocode.}

\item{lon}{A column in \code{df} containing the longitude to geocode.}

\item{type}{A string indicating the type of geocoding to be performed.
Must be one of "position", "address", or "matchConfidence".}

\item{.test_batch}{A logical. If \code{TRUE}, the function will run in test mode,
processing only a small batches of the data.}
}
\value{
A data frame containing the original data plus geocoding results.
}
\description{
This function interacts with the AZM geocoding service to
reverse geocode coordinates in a data frame, handling batching and type selection.
}
