% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand.R
\docType{data}
\name{cand}
\alias{cand}
\title{Candidate Addresses}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{\code{id}}{A character vector of candidate identifiers.}
\item{\code{address}}{A character vector of candidate addresses.}
}
}
\source{
\url{[URL where the data was taken from, if applicable]}
}
\usage{
cand
}
\description{
A dataset containing IDs and addresses of certain candidates.
}
\examples{
data(cand)
head(cand)
}
\keyword{datasets}
