% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coords}
\alias{coords}
\title{Coordinates dataset}
\format{
A data frame with 6 rows and 2 variables:
\describe{
\item{lat}{Latitude of the location. Numeric.}
\item{lon}{Longitude of the location. Numeric.}
}
}
\usage{
coords
}
\description{
A dataset containing latitude and longitude values for certain locations.
}
\examples{
head(coords)

# Plotting the coordinates
plot(coords$lon, coords$lat)

}
\keyword{datasets}
