% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/az_url.R
\name{get_az_url}
\alias{get_az_url}
\title{Retrieve Full Azure Maps API URL}
\usage{
get_az_url()
}
\value{
A string, the complete API URL including the subscription key.
}
\description{
This function retrieves and returns the full URL for the Azure Maps API,
ensuring the necessary API key is available in the user's environment.
}
\examples{
\dontrun{
url <- get_az_url()
}
}
