% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azm_url.R
\name{get_azm_url}
\alias{get_azm_url}
\title{Retrieve Full Azure Maps API URL}
\usage{
get_azm_url(direction)
}
\arguments{
\item{direction}{Forward "fwd" or Reverse "rev"}
}
\value{
A string, the complete API URL including the subscription key.
}
\description{
This function retrieves and returns the full URL for the Azure Maps API,
ensuring the necessary API key is available in the user's environment.
}
\examples{
\dontrun{
url <- get_azm_url()
}
}
