% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{req_body_cond}
\alias{req_body_cond}
\title{Conditionally Set Request Body}
\usage{
req_body_cond(request, batch_json)
}
\arguments{
\item{request}{A request object created by httr::request().}

\item{batch_json}{Either a JSON object or a character string representing
a file path to a JSON file.}
}
\value{
The modified request object.
}
\description{
This function conditionally sets the body of an HTTP request depending on
the type of \code{batch_json}. If \code{batch_json} is a JSON object, \code{req_body_raw}
is used; if it's a valid file path (character string), \code{req_body_file} is used.
}
